requirements.txt
PyMuPDF>=1.23.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
